pgconfig: fix error in resourcestore moving a path to itself

+++ src/catalog/backends/pgconfig/src/main/java/org/geoserver/cloud/backend/pgconfig/resource/PgsqlResourceStore.java
@@ -275,6 +275,10 @@ public class PgsqlResourceStore implements ResourceStore {
         if (!source.exists()) {
             return false;
         }
+        if (source.path().equals(target.path())) {
+            return true;
+        }
+
         if (target.exists()) {
             target.delete();
         }
